Split object provenance from rooted transforms
Separate spawned and referenced prim behavior from root-transform state handling
while preserving the existing Object and ObjectReference API names.

BREAKING CHANGE:
- Direct ObjectBase subclasses must implement get_object_pose
and set_object_pose, or compose RootedTransform.
- ObjectBase.set_prim_path() has been removed. Prim paths must
now be supplied during object construction and remain immutable afterward.
